+++ frontend/src/pages/auth/Registration.tsx
