+++ frontend/src/pages/game/Canvas.tsx
