+++ frontend/src/pages/game/GameEnd.tsx
