+++ frontend/src/pages/game/GameStart.tsx
