+++ frontend/src/pages/profile/StatsCard.tsx
